Move dialog/conversation into a separate "bot" directory.
* These will eventually become part of a larger "bot" abstraction, that
  can be broken out into a separate module.
