+++ Gruntfile.js
@@ -15,9 +15,7 @@ module.exports = function(grunt) {
           'src/index.js',
           'src/lib/args.js',
           'src/lib/bot.js',
-          'src/lib/conversation.js',
-          'src/lib/db.js',
-          'src/lib/dialog.js',
+          'src/lib/bot/**/*.js',
           'src/lib/slack.js',
           'src/thanksbot/**',
           'src/robocoup/index.js',
