+++ src/lib/bot/conversation.js
