+++ src/lib/bot/dialog.js
